move fee asset parsing into a separate entity
