+++ novawallet/Common/Model/StatemineAsset/StatemineAssetSerializer.swift
@@ -4,7 +4,6 @@ import BigInt
 
 enum StatemineAssetSerializerError: Error {
     case assetIdTypeNotFound(palletName: String?)
-    case feeAssetIdTypeNotFound
 }
 
 enum StatemineAssetSerializer {
@@ -21,21 +20,6 @@ enum StatemineAssetSerializer {
         return call.arguments[0].type
     }
 
-    private static func extractFeeAssetIdType(from codingFactory: RuntimeCoderFactoryProtocol) -> String? {
-        guard
-            let extensionType = codingFactory.metadata.getSignedExtensionType(
-                for: Extrinsic.SignedExtensionId.assetTxPayment
-            ),
-            let extensionTypeNode = codingFactory.getTypeNode(for: extensionType) as? StructNode,
-            let assetIdType = extensionTypeNode.typeMapping.first(
-                where: { $0.name == "assetId" }
-            )?.node.typeName else {
-            return nil
-        }
-
-        return assetIdType
-    }
-
     static func decode(
         assetId: String,
         palletName: String?,
@@ -48,26 +32,6 @@ enum StatemineAssetSerializer {
 
         guard let assetIdType = extractAssetIdType(from: codingFactory, palletName: palletName) else {
             throw StatemineAssetSerializerError.assetIdTypeNotFound(palletName: palletName)
-        }
-
-        let data = try Data(hexString: assetId)
-
-        let decoder = try codingFactory.createDecoder(from: data)
-
-        return try decoder.read(type: assetIdType)
-    }
-
-    static func decodeFeeAssetId(
-        _ assetId: String,
-        codingFactory: RuntimeCoderFactoryProtocol
-    ) throws -> JSON {
-        // assetId is either integer or complicated data structure
-        guard assetId.isHex() else {
-            return JSON.stringValue(assetId)
-        }
-
-        guard let assetIdType = extractFeeAssetIdType(from: codingFactory) else {
-            throw StatemineAssetSerializerError.feeAssetIdTypeNotFound
         }
 
         let data = try Data(hexString: assetId)
